==> Scripts/1. Data Import & Descriptive Analysis.ipynb <==
#01. Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

#02. Data Types

x = 2

x

print(x)

type(x)

y = 5.678

type(y)

z = 'Brazil'

type(z)

type('Mania')

type(True)

type(False)

x + 4

x + y

x = x + 4

print(x)

a = 'Hello '

b = 'Miguel'

a + b

n = 'potatoes'

n * 10

x == y

type(x == y)

#03. Exercise 4.2

# 7
80 + 20

# 7
150 - 50

# 7
30 + 70

# 8
t = 200

# 8
r = 2

# 8
t / r

# 9
p = 'Port'

# 9
g = 'ugal'

# 9
p + g

# 10
l = 'Learning Python is fun!'

#10
i = ' It is a valuable skill.'

#10
l + i

# Bonus
h = 'Miguel '

# Bonus
h * 100
