==> Scripts/2. Data Wrangling & Subsetting.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

# Orders DF
df = pd.read_csv(r'/Users/miguel/Documents/Instacart Basket Analysis/02 Data/Original Data/orders.csv', index_col = False)

path = r'/Users/miguel/Documents/Instacart Basket Analysis'

path

df.head()

df.head(30)

df.tail()

df.shape

df.info()

df.dtypes

df.columns

df.describe()

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

print(path)

df.describe()

# 5
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# 5
vars_list

# 5
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

# 6
df.head()

df.tail()

# 7
df_prods = pd.read_csv(r'/Users/miguel/Documents/Instacart Basket Analysis/02 Data/Original Data/products.csv', index_col = False)

# 8
df_prods.head(20)

# 8
df_prods.tail(35)

# 9
df_prods.columns

# 10
df_prods.shape

# 11
df_prods.describe()

# 11
df_prods.dtypes

# 11
Aisle_id Max = 134
Department_id Data Type = Integer
