==> Scripts/3. Data Consistency Checks.ipynb <==
# Data Wrangling Reading Portion

# Importing libraries
import pandas as pd
import numpy as np
import os

# Importing orders file
df_orders = pd.read_csv(r'/Users/miguel/Documents/Instacart Basket Analysis/02 Data/Original Data/orders.csv', index_col = False)

# Importing products file
df_products = pd.read_csv(r'/Users/miguel/Documents/Instacart Basket Analysis/02 Data/Original Data/products.csv', index_col = False)

# Visually dropping 'eval_set' column from data frame
df_orders.drop(columns = ['eval_set'])

# Finding missing values in a column
df_orders['days_since_prior_order'].value_counts(dropna = False)

# Renaming columns
df_orders.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Checking new column name
df_orders.head()

# Changing columns data type
df_orders['order_id'] = df_orders['order_id'].astype('str')

# Checking new column data type
df_orders['order_id'].dtype

# Importing departments file
df_departments = pd.read_csv(r'/Users/miguel/Documents/Instacart Basket Analysis/02 Data/Original Data/departments.csv', index_col = False)

# Checking departments file
df_departments.head()

# Transposing departments file
df_departments.T

# Permanently transposing departments file
df_departments_t = df_departments.T

# Adding an index to departments file
df_departments_t.reset_index()

# Selecting first row of departments file for header
new_header = df_departments_t.iloc[0]

# Checking work
new_header

# New df for rows beyond the first row
df_departments_t_new = df_departments_t[1:]

# Checking work
df_departments_t_new

# Changing header to original column names
df_departments_t_new.columns = new_header

# Checking work
df_departments_t_new

# Turning df into data dictionary
data_dict = df_departments_t_new.to_dict('index')

# Checking work
data_dict

df_products.head()

# Using data dictionary to read table above
print(data_dict.get('19'))

# Creating new subset of data for snacks
df_snacks =  df_products[df_products['department_id']==19]

# Checking work
df_snacks.head()

# Second way of creating subset of data for snacks
df_snacks_2 = df_products.loc[df_products['department_id'].isin([19])]

# Checking work
df_snacks_2.head()

# Task 4.4

df_orders.head()

# 2 identifying and changing data type for specific column: user_id
df_orders['user_id'] = df_orders['user_id'].astype('str')

# 2 Checking work
df_orders['user_id'].dtype

# 3 Changing column name: order_hour_of_day
df_orders.rename(columns = {'order_hour_of_day' : 'order_time'}, inplace = True)

# 3 Checking work
df_orders.head()

# 4 Busiest hour for placing orders: frequency of order time
df_orders['order_time'].value_counts(dropna = False)

The busiest hour is 10 AM, as it is the hour of the day with the highest frequency of orders (as shown above).

# 5 Meaning of value '4' in the department_id column
print(data_dict.get('4'))

# 6 Subset of data for breakfast items
df_breakfast =  df_products[df_products['department_id']==14]

# 6 Checking work
df_breakfast

# 7 Subset of data for customers that like to throw dinner parties
df_dinner_parties = df_products.loc[df_products['department_id'].isin([5,20,7,12])]

# 7 Checking work
df_dinner_parties

#8 The dataframe created above has 7650 rows and 5 columns.

# 9 Finding information about user_id 1
df_user_id_1 =  df_orders[df_orders['user_id']=='1']

# 9 Checking work
df_user_id_1

#10 Some of basic stats and information we can gather from user 1, is that he/she made 11 orders in total, and that he/she usually orders from Instacart on day of the week 4, even though it varies a lot from the 11 times he/she has ordered.

# Final steps before exporting
df_orders = df_orders.drop(columns = ['eval_set'])

# Check
df_orders.head()

# Final steps before exporting
path = r'/Users/miguel/Documents/Instacart Basket Analysis'

# Exporting df_orders
df_orders.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

# Exporting df_departments_t_new
df_departments_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
