==> Scripts/4. Combining & Exporting Data.ipynb <==
# Data Consistency Reading

# Importing libraries
import pandas as pd
import numpy as np
import os

# Defining path to import files
path = r'/Users/miguel/Documents/Instacart Basket Analysis'

# Importing orders wranged file
df_orders = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# Importing products file
df_products = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Create a dataframe
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a','b',1, True]

df_test.head()

# Checking for mixed types
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# Assigning data type to mixed column
df_test['mix'] = df_test['mix'].astype('str')

# Finding missing values in products file
df_products.isnull().sum()

# Creating product_name missing values subset
df_nan = df_products[df_products['product_name'].isnull() == True]

# Checking work
df_nan

# Comparing missing values to overall dataframe
df_products.shape

# Removing 16 missing values from products dataframe
df_products_clean = df_products[df_products['product_name'].isnull() == False]

# Checking work
df_products_clean.shape

# Looking for full duplicates in clean products dataframe
df_dups = df_products_clean[df_products_clean.duplicated()]

# Checking work
df_dups

# Checking number of rows before removing duplicates
df_products_clean.shape

# Creating new dataframe removing duplicates
df_products_clean_no_dups = df_products_clean.drop_duplicates()

# Checking work
df_products_clean_no_dups.shape

# Task 4.5

# 2 Finding data inconsistencies in the orders file
df_orders.describe()

#2 Order time max should be 24, representing the 24 hours in a day.

# 3 Checking for mixed type data
for col in df_orders.columns.tolist():
  weird = (df_orders[[col]].map(type) != df_orders[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_orders[weird]) > 0:
    print (col)

#4 No mixed type data was found in the orders dataframe

# 5 Checking for missing values in orders dataframe
df_orders.isnull().sum()

#5 There were 206,209 missing values in the 'days_since_prior_order' column. Since this is such a significant number of rows woth missing values, I believe that they are missing by design. This means there were 206,209 times where it was the first time that customers were ordering something from Instacart. 

#6 I decided to leave the missing values because they are significant for the analysis, as they represent customers making their first purchase at Instacart.

# 7 Checking for duplicates in the orders dataframe
df_orders_dups = df_orders[df_orders.duplicated()]

# 7 Check work
df_orders_dups

#7 No duplicate values were found in the orders dataframe

#8 No duplicate values were found in the orders dataframe

df_orders.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'))

df_products.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))
