==> Scripts/5. Deriving New Variables Pt1.ipynb <==
# Data Immersion 4_6

# Import libraries

import pandas as pd
import numpy as np
import os

path = r'/Users/miguel/Documents/Instacart Basket Analysis'

# Importing orders_products_prior file
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

# Checking output
df_ords_prior.head()

# Checking shape
df_ords_prior.shape

# Importing orders file
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col = False)

# Checking shape
df_ords.shape

# Merging ords and ords_prior dataframes
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# Checking output
df_merged_large.head()

# Checking for full match
df_merged_large['_merge'].value_counts()

# Dropping unnecessary columns
df_merged_large = df_merged_large.drop(columns = ['Unnamed: 0.1'])

# Checking work
df_merged_large.head()

# Dropping unnecessary columns
df_merged_large = df_merged_large.drop(columns = ['Unnamed: 0'])

# Checking work
df_merged_large.head()

# Exporting new merged dataframe in pickle format
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
