==> Scripts/6. Deriving New Variables Pt2.ipynb <==
# Task 4.6

# Import libraries

import pandas as pd
import numpy as np
import os

path = r'/Users/miguel/Documents/Instacart Basket Analysis'

# Importing orders_products_combined file
df_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# Checking shape
df_merged.shape

# Importing clean products file
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = False)

# Checking for common columns
df_merged.head()

# Checking for common columns
df_prods.head()

# Merging the 2 dataframes
df_ords_prods_merged = df_merged.merge(df_prods, on = 'product_id', indicator = "merge_flag")

# Checking output
df_ords_prods_merged.head()

# Checking merge flag
df_ords_prods_merged["merge_flag"].value_counts()

# Dropping unnecessary columns
df_ords_prods_merged = df_ords_prods_merged.drop(columns = ['Unnamed: 0'])

# Checking work
df_ords_prods_merged.head()

# Dropping unnecessary columns
df_ords_prods_merged = df_ords_prods_merged.drop(columns = ['_merge'])

# Checking work
df_ords_prods_merged.head()

# Dropping unnecessary columns
df_ords_prods_merged = df_ords_prods_merged.drop(columns = ['merge_flag'])

# Checking work
df_ords_prods_merged.head()

# Exporting merged dataframe
df_ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
