==> Scripts/7. Grouping Data & Aggregating Variables.ipynb <==
# 4.7 Reading

# Importing libraries
import pandas as pd
import numpy as np
import os

path = r'/Users/miguel/Documents/Instacart Basket Analysis'

# Importing merged file
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

# Creating subset for first million rows of data
df = df_ords_prods_merge[:1000000]

# Define a function
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Apply the function
df['price_range'] = df.apply(price_label, axis=1)

# Checking work
df['price_range'].value_counts(dropna = False)

# Checking most expensive product
df['prices'].max()

# Creating new variable using loc() function
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Creating new variable using loc() function
df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

# Creating new variable using loc() function
df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Checking work
df['price_range_loc'].value_counts(dropna = False)

# Applying new variable to whole dataframe
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Applying new variable to whole dataframe
df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

# Applying new variable to whole dataframe
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Checking work
df_ords_prods_merge['price_range_loc'].value_counts(dropna = False)

# Printing the frequency of the “orders_day_of_week” column
df_ords_prods_merge['orders_day_of_week'].value_counts(dropna = False)

# Creating new column to show busiest days

result = []

for value in df_ords_prods_merge['orders_day_of_week']:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# Checking output
result

# Adding new column to dataframe
df_ords_prods_merge['Busiest_Day'] = result

# Checking output
df_ords_prods_merge['Busiest_Day'].value_counts(dropna = False)

# Task 4.7

# 2: Creating "Busiest Days" column

result = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest days")
  elif value == 1:
    result.append("Busiest days")
  elif value == 4:
    result.append("Slow days")
  elif value == 3:
    result.append("Slow days")
  else:
    result.append("Regularly busy")

# 2: Checking output
result

# 2: Adding new column to dataframe
df_ords_prods_merge['Busiest_Days'] = result

# 3: Checking the values of the new column
df_ords_prods_merge['Busiest_Days'].value_counts(dropna = False)

#3: As expected, there are more regularly busy days than really busy days and slow days. The positive we takeaway from the data is that there are more busy days than there are slow days, which means Instacart is selling at a good rate. 

# 4: Identifying busiest hours of day for orders
df_ords_prods_merge['order_time'].value_counts(dropna = False)

# 4: Creating "Busiest Period of Day" column

result = []

for value in df_ords_prods_merge['order_time']:
  if value in [10,11,14,15,13,12,16,9]:
    result.append("Most orders")
  elif value in [3,4,2,5,1,0,6,23]:
    result.append("Fewest orders")
  else:
    result.append("Average orders")

# 4: Checking output
result

# 4: Adding new column to dataframe
df_ords_prods_merge['Busiest_Period_of_Day'] = result

# 5: Printing frequency for values in new column
df_ords_prods_merge['Busiest_Period_of_Day'].value_counts(dropna = False)

# 7: Exporting dataframe with new columns
df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_new_variables.pkl'))
