==> Scripts/9. Visualizations Pt 1.ipynb <==
# Exercise 4.9 Part 1

#3: Importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

path = r'/Users/miguel/Documents/Instacart Basket Analysis'

#3: Importing file
df_customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col = False)

#4: Checking for errors and inconsistencies
df_customers.head()

#4: Checking for errors and inconsistencies
df_customers.describe()

#4: Checking for errors and inconsistencies
df_customers.info

#4: Changing column names 
df_customers.rename(columns = {'Surnam' : 'Surname'}, inplace = True)
df_customers.rename(columns = {'fam_status' : 'Marital_Status'}, inplace = True)
df_customers.rename(columns = {'n_dependants' : 'Number_of_Dependants'}, inplace = True)
df_customers.rename(columns = {'STATE': 'State'}, inplace = True)
df_customers.rename(columns = {'user_id': 'User_ID'}, inplace = True)
df_customers.rename(columns = {'date_joined': 'Date_Joined'}, inplace = True)
df_customers.rename(columns = {'income': 'Income'}, inplace = True)

#4: Checking work
df_customers.head()

#5: Checking for missing values
df_customers['First Name'].value_counts(dropna = False)

#5: Checking for missing values
df_customers['Surname'].value_counts(dropna = False)

#5: Checking for missing values
df_customers['Gender'].value_counts(dropna = False)

#5: Checking for missing values
df_customers['State'].value_counts(dropna = False)

#5: Checking for missing values
df_customers['Age'].value_counts(dropna = False)

#5: Checking for missing values
df_customers['Date_Joined'].value_counts(dropna = False)

#5: Checking for missing values
df_customers['number_of_dependants'].value_counts(dropna = False)

#5: Checking for missing values
df_customers['Marital_Status'].value_counts(dropna = False)

#5: Checking for missing values
df_customers['Income'].value_counts(dropna = False)

#5: Addressing missing values
df_customers['First Name'].fillna('Unknown', inplace = True)

#5: Checking work
df_customers['First Name'].value_counts(dropna = False)

df_customers.shape

#5: Checking for duplicates
df_customers_dups = df_customers[df_customers.duplicated()]

#5: No duplicates were found

#5: Checking work
df_customers_dups

#5: Checking for mixed data types
for col in df_customers.columns.tolist():
    weird = (df_customers[[col]].map(type) != df_customers[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_customers[weird]) > 0:
        print (col)

#5: No mixed data types were found

#6: Importing file
df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_updated_4_8.pkl'))

#6: Checking key columns
df_ords_prods.head()

#6: Renaming df_customers User_ID column for effective merge
df_customers.rename(columns = {'User_ID': 'user_id'}, inplace = True)

#6: Checking work
df_customers.head()

#6: Merging customers df with prepared Instacart data
df_ords_prods_customer = df_ords_prods.merge(df_customers, on = ['user_id'])

#6: Checking work
df_ords_prods_customer.head()

#8: Exporting new df
df_ords_prods_customer.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_customers_4_9.pkl'))
