==> Scripts/11. Final Results & Visuals Pt1.ipynb <==
# Exercise 4.10

#1: Importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

path = r'/Users/miguel/Documents/Instacart Basket Analysis'

#1: Importing file
df_ords_prods_customer = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_customers_4_9.pkl'))

df_ords_prods_customer.head()

#2: Addressing PII by droping name columns
df_ords_prods_customer = df_ords_prods_customer.drop(columns = ['First Name', 'Surname'])

#3: Creating region column

result = []

for value in df_ords_prods_customer['State']:
  if value in ['Maine','New Hampshire','Vermont','Massachusetts','Rhode Island','Connecticut','New York','Pennsylvania','New Jersey']:
    result.append("Northeast")
  elif value in ['Wisconsin','Michigan','Illinois','Indiana','Ohio','North Dakota','South Dakota','Nebraska','Kansas','Minnesota','Iowa','Missouri']:
    result.append("Midwest")
  elif value in ['Delaware','Maryland','District of Columbia','Virginia','West Virginia','North Carolina','South Carolina','Georgia','Florida','Kentucky','Tennessee','Mississippi','Alabama','Oklahoma','Texas','Arkansas','Louisiana']:
    result.append("South")
  else:
    result.append("West")

#3: Adding new column to df
df_ords_prods_customer['Region'] = result

#3: Checking work
df_ords_prods_customer.head()

#3: Checking work
df_ords_prods_customer['Region'].value_counts(dropna = False)

#3: Create a crosstab to find difference in spending habits between regions
crosstab = pd.crosstab(df_ords_prods_customer['spending_flag'], df_ords_prods_customer['Region'], dropna = False)

#3: Copy crosstab
crosstab.to_clipboard()

#3: Paste crosstab
crosstab

#4: Create an exclusion flag
df_ords_prods_customer.loc[df_ords_prods_customer['max_order'] < 5, 'activity_flag'] = 'Low activity'

#4:  Create an exclusion flag
df_ords_prods_customer.loc[df_ords_prods_customer['max_order'] >= 5, 'activity_flag'] = 'Normal activity'

#4: Checking work
df_ords_prods_customer['activity_flag'].value_counts(dropna = False)

#4: Creating a subset for low activity customers
df_low_activity = df_ords_prods_customer[df_ords_prods_customer['activity_flag'] =='Low activity']

#4: Checking work
df_low_activity.head(15)

#4: Creating a subset for normal activity customers
df_normal_activity = df_ords_prods_customer[df_ords_prods_customer['activity_flag'] =='Normal activity']

#4: Checking work
df_normal_activity['activity_flag'].value_counts(dropna = False)

#4: Exporting dataframe
df_ords_prods_customer.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'df_4_10.pkl'))

#4: Exporting normal activity dataframe
df_normal_activity.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'df_normal_activity.pkl'))

#4: Exporting low activity dataframe
df_low_activity.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'df_low_activity.pkl'))
